+++ spec/integration/rest/conversations/v1/conversation.spec.js
@@ -64,6 +64,10 @@ describe('Conversation', function() {
           'date_created': '2015-12-16T22:18:37Z',
           'date_updated': '2015-12-16T22:18:38Z',
           'state': 'inactive',
+          'timers': {
+              'date_inactive': '2015-12-16T22:19:38Z',
+              'date_closed': '2015-12-16T22:28:38Z'
+          },
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -119,6 +123,10 @@ describe('Conversation', function() {
           'date_created': '2015-12-16T22:18:37Z',
           'date_updated': '2015-12-16T22:18:38Z',
           'state': 'inactive',
+          'timers': {
+              'date_inactive': '2015-12-16T22:19:38Z',
+              'date_closed': '2015-12-16T22:28:38Z'
+          },
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -210,6 +218,10 @@ describe('Conversation', function() {
           'date_created': '2015-12-16T22:18:37Z',
           'date_updated': '2015-12-16T22:18:38Z',
           'state': 'active',
+          'timers': {
+              'date_inactive': '2015-12-16T22:19:38Z',
+              'date_closed': '2015-12-16T22:28:38Z'
+          },
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -243,6 +255,10 @@ describe('Conversation', function() {
                   'date_created': '2015-12-16T22:18:37Z',
                   'date_updated': '2015-12-16T22:18:38Z',
                   'state': 'active',
+                  'timers': {
+                      'date_inactive': '2015-12-16T22:19:38Z',
+                      'date_closed': '2015-12-16T22:28:38Z'
+                  },
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -279,6 +295,10 @@ describe('Conversation', function() {
                   'date_created': '2015-12-16T22:18:37Z',
                   'date_updated': '2015-12-16T22:18:38Z',
                   'state': 'active',
+                  'timers': {
+                      'date_inactive': '2015-12-16T22:19:38Z',
+                      'date_closed': '2015-12-16T22:28:38Z'
+                  },
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -320,6 +340,10 @@ describe('Conversation', function() {
                   'date_created': '2015-12-16T22:18:37Z',
                   'date_updated': '2015-12-16T22:18:38Z',
                   'state': 'active',
+                  'timers': {
+                      'date_inactive': '2015-12-16T22:19:38Z',
+                      'date_closed': '2015-12-16T22:28:38Z'
+                  },
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',
@@ -376,6 +400,10 @@ describe('Conversation', function() {
                   'date_created': '2015-12-16T22:18:37Z',
                   'date_updated': '2015-12-16T22:18:38Z',
                   'state': 'active',
+                  'timers': {
+                      'date_inactive': '2015-12-16T22:19:38Z',
+                      'date_closed': '2015-12-16T22:28:38Z'
+                  },
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'participants': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants',